requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
Flask-Login
werkzeug
python-dotenv
